+++ variables.tf
@@ -74,7 +74,7 @@ variable "vnet_config" {
 
 variable "skip_provider_registration" {
   type        = bool
-  description = "Allows you to skip the provider registration when initilizing the azurerm provider. This is useful in development environments where not every provider can be registered."
+  description = "Allows you to skip the provider registration when initilizing the azurerm provider in this configuration and the created configuration. This is useful in development environments where not every provider can be registered."
   default     = false
 }