+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- rename `vnet.tf` to `network.tf` as it contains Resource Group for `Network`
+- remove `locals` from `network.tf`/`vnet.tf`
+
 ## [1.0.1] - 2024-03-25
 
 ### Changed