+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Set the `skip_provider_registration` of `azurerm` provider in created terraform configuration to `var.skip_provider_registration` instead of `true`
 
+### Changed
+
+- rename `vnet.tf` to `network.tf` as it contains Resource Group for `Network`
+- remove `locals` from `network.tf`/`vnet.tf`
+
 ## [1.0.1] - 2024-03-25
 
 ### Changed